requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
astroid==2.9.1
backports.zoneinfo==0.2.1
black==21.8b0
click==8.0.1
colorama==0.4.4
cycler==0.11.0
Django==4.0
django-crispy-forms==1.13.0
django-easy-pdf==0.1.1
django-excel-response==2.0.5
django-reportengine==0.3.1
django-slick-reporting==0.6.3
et-xmlfile==1.1.0
fonttools==4.28.5
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.2
inflect==5.3.0
isort==5.10.1
jdcal==1.4.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.0
openpyxl==2.4.11
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.3.0
pylint==2.12.2
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
regex==2021.8.28
reportlab==3.6.5
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typed-ast==1.5.1
typing-extensions==3.10.0.2
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
XlsxWriter==3.0.2
